edits 2017/2/22
1. fixed contrast level text file read-in so empty strings are not
converted to floats (throws an error)
2. reorganized process_images() loop through paths and contrast levels,
should cut processing time.
3. TODO: (ln63) change all background pixels to transparent. We don't
have to do this if we load in bmp files w/out alpha layers or if we load
in jpg or tif files w/ alpha layers - either would work, and if we need
to save bmp w/ alpha layers, we can.