adding BMPA functionality
1. changed string formatting and added nicer output text
2. moved file opening and converting to RGBA within contrast editing
loop. without this, each step of contrast editing edited the previous
step, leading to things like 50% of 40% of 80% and so on.
3.  IMPORTANT STEP! Added imgdraw.floodfill() method to contrast level
loop. This samples a pixel and floodfills the area adjacent to that
pixel with the color of your choosing! This edits the image object and
returns nothing. Easy peasy.

NOTE: You still can't use BMPA files as input to this script. However,
you can use any other image files as input (including bmp) and they will
output as bmp files. ALWAYS!

I can add a log function (output log text of procedures) if you want. Do
you want?